Fix build

+++ test/e2e/storage/volumes.go
@@ -51,10 +51,9 @@ import (
 	. "github.com/onsi/gomega"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/kubernetes/pkg/client/clientset_generated/clientset"
 	"k8s.io/kubernetes/pkg/api/v1"
+	"k8s.io/kubernetes/pkg/client/clientset_generated/clientset"
 	"k8s.io/kubernetes/pkg/cloudprovider/providers/vsphere"
-	kubeletapis "k8s.io/kubernetes/pkg/kubelet/apis"
 	"k8s.io/kubernetes/test/e2e/framework"
 )
 
@@ -80,11 +79,13 @@ func DeleteCinderVolume(name string) error {
 }
 
 // These tests need privileged containers, which are disabled by default.
-var _ = framework.KubeDescribe("Volumes", func() {
+var _ = framework.KubeDescribe("Volumes [Volume]", func() {
 	f := framework.NewDefaultFramework("volume")
 
+	// If 'false', the test won't clear its volumes upon completion. Useful for debugging,
 	// note that namespace deletion is handled by delete-namespace flag
-	// filled inside BeforeEach
+	clean := true
+	// filled in BeforeEach
 	var cs clientset.Interface
 	var namespace *v1.Namespace
 
@@ -97,10 +98,23 @@ var _ = framework.KubeDescribe("Volumes", func() {
 	// NFS
 	////////////////////////////////////////////////////////////////////////
 
-	Describe("NFS", func() {
+	framework.KubeDescribe("NFS", func() {
 		It("should be mountable", func() {
-			config, _, serverIP := framework.NewNFSServer(cs, namespace.Name, []string{})
-			defer framework.VolumeTestCleanup(f, config)
+			config := framework.VolumeTestConfig{
+				Namespace:   namespace.Name,
+				Prefix:      "nfs",
+				ServerImage: framework.NfsServerImage,
+				ServerPorts: []int{2049},
+			}
+
+			defer func() {
+				if clean {
+					framework.VolumeTestCleanup(f, config)
+				}
+			}()
+			pod := framework.StartVolumeServer(cs, config)
+			serverIP := pod.Status.PodIP
+			framework.Logf("NFS server IP address: %v", serverIP)
 
 			tests := []framework.VolumeTest{
 				{
@@ -124,25 +138,71 @@ var _ = framework.KubeDescribe("Volumes", func() {
 	// Gluster
 	////////////////////////////////////////////////////////////////////////
 
-	Describe("GlusterFS", func() {
+	framework.KubeDescribe("GlusterFS [Feature:Volumes]", func() {
 		It("should be mountable", func() {
 			//TODO (copejon) GFS is not supported on debian image.
-			framework.SkipUnlessNodeOSDistroIs("gci", "ubuntu")
+			framework.SkipUnlessNodeOSDistroIs("gci")
 
-			// create gluster server and endpoints
-			config, _, _ := framework.NewGlusterfsServer(cs, namespace.Name)
-			name := config.Prefix + "-server"
+			config := framework.VolumeTestConfig{
+				Namespace:   namespace.Name,
+				Prefix:      "gluster",
+				ServerImage: framework.GlusterfsServerImage,
+				ServerPorts: []int{24007, 24008, 49152},
+			}
+
 			defer func() {
-				framework.VolumeTestCleanup(f, config)
-				err := cs.CoreV1().Endpoints(namespace.Name).Delete(name, nil)
-				Expect(err).NotTo(HaveOccurred(), "defer: Gluster delete endpoints failed")
+				if clean {
+					framework.VolumeTestCleanup(f, config)
+				}
 			}()
+			pod := framework.StartVolumeServer(cs, config)
+			serverIP := pod.Status.PodIP
+			framework.Logf("Gluster server IP address: %v", serverIP)
+
+			// create Endpoints for the server
+			endpoints := v1.Endpoints{
+				TypeMeta: metav1.TypeMeta{
+					Kind:       "Endpoints",
+					APIVersion: "v1",
+				},
+				ObjectMeta: metav1.ObjectMeta{
+					Name: config.Prefix + "-server",
+				},
+				Subsets: []v1.EndpointSubset{
+					{
+						Addresses: []v1.EndpointAddress{
+							{
+								IP: serverIP,
+							},
+						},
+						Ports: []v1.EndpointPort{
+							{
+								Name:     "gluster",
+								Port:     24007,
+								Protocol: v1.ProtocolTCP,
+							},
+						},
+					},
+				},
+			}
+
+			endClient := cs.Core().Endpoints(config.Namespace)
+
+			defer func() {
+				if clean {
+					endClient.Delete(config.Prefix+"-server", nil)
+				}
+			}()
+
+			if _, err := endClient.Create(&endpoints); err != nil {
+				framework.Failf("Failed to create endpoints for Gluster server: %v", err)
+			}
 
 			tests := []framework.VolumeTest{
 				{
 					Volume: v1.VolumeSource{
 						Glusterfs: &v1.GlusterfsVolumeSource{
-							EndpointsName: name,
+							EndpointsName: config.Prefix + "-server",
 							// 'test_vol' comes from test/images/volumes-tester/gluster/run_gluster.sh
 							Path:     "test_vol",
 							ReadOnly: true,
@@ -166,10 +226,27 @@ var _ = framework.KubeDescribe("Volumes", func() {
 	// are installed on all nodes!
 	// Run the test with "go run hack/e2e.go ... --ginkgo.focus=iSCSI"
 
-	Describe("iSCSI [Feature:Volumes]", func() {
+	framework.KubeDescribe("iSCSI [Feature:Volumes]", func() {
 		It("should be mountable", func() {
-			config, _, serverIP := framework.NewISCSIServer(cs, namespace.Name)
-			defer framework.VolumeTestCleanup(f, config)
+			config := framework.VolumeTestConfig{
+				Namespace:   namespace.Name,
+				Prefix:      "iscsi",
+				ServerImage: framework.IscsiServerImage,
+				ServerPorts: []int{3260},
+				ServerVolumes: map[string]string{
+					// iSCSI container needs to insert modules from the host
+					"/lib/modules": "/lib/modules",
+				},
+			}
+
+			defer func() {
+				if clean {
+					framework.VolumeTestCleanup(f, config)
+				}
+			}()
+			pod := framework.StartVolumeServer(cs, config)
+			serverIP := pod.Status.PodIP
+			framework.Logf("iSCSI server IP address: %v", serverIP)
 
 			tests := []framework.VolumeTest{
 				{
@@ -196,10 +273,28 @@ var _ = framework.KubeDescribe("Volumes", func() {
 	// Ceph RBD
 	////////////////////////////////////////////////////////////////////////
 
-	Describe("Ceph RBD [Feature:Volumes]", func() {
+	framework.KubeDescribe("Ceph RBD [Feature:Volumes]", func() {
 		It("should be mountable", func() {
-			config, _, serverIP := framework.NewRBDServer(cs, namespace.Name)
-			defer framework.VolumeTestCleanup(f, config)
+			config := framework.VolumeTestConfig{
+				Namespace:   namespace.Name,
+				Prefix:      "rbd",
+				ServerImage: framework.RbdServerImage,
+				ServerPorts: []int{6789},
+				ServerVolumes: map[string]string{
+					// iSCSI container needs to insert modules from the host
+					"/lib/modules": "/lib/modules",
+					"/sys":         "/sys",
+				},
+			}
+
+			defer func() {
+				if clean {
+					framework.VolumeTestCleanup(f, config)
+				}
+			}()
+			pod := framework.StartVolumeServer(cs, config)
+			serverIP := pod.Status.PodIP
+			framework.Logf("Ceph server IP address: %v", serverIP)
 
 			// create secrets for the server
 			secret := v1.Secret{
@@ -217,10 +312,12 @@ var _ = framework.KubeDescribe("Volumes", func() {
 				Type: "kubernetes.io/rbd",
 			}
 
-			secClient := cs.CoreV1().Secrets(config.Namespace)
+			secClient := cs.Core().Secrets(config.Namespace)
 
 			defer func() {
-				secClient.Delete(config.Prefix+"-secret", nil)
+				if clean {
+					secClient.Delete(config.Prefix+"-secret", nil)
+				}
 			}()
 
 			if _, err := secClient.Create(&secret); err != nil {
@@ -250,11 +347,11 @@ var _ = framework.KubeDescribe("Volumes", func() {
 			framework.TestVolumeClient(cs, config, &fsGroup, tests)
 		})
 	})
-
 	////////////////////////////////////////////////////////////////////////
 	// Ceph
 	////////////////////////////////////////////////////////////////////////
-	Describe("CephFS [Feature:Volumes]", func() {
+
+	framework.KubeDescribe("CephFS [Feature:Volumes]", func() {
 		It("should be mountable", func() {
 			config := framework.VolumeTestConfig{
 				Namespace:   namespace.Name,
@@ -263,8 +360,14 @@ var _ = framework.KubeDescribe("Volumes", func() {
 				ServerPorts: []int{6789},
 			}
 
-			defer framework.VolumeTestCleanup(f, config)
-			_, serverIP := framework.CreateStorageServer(cs, config)
+			defer func() {
+				if clean {
+					framework.VolumeTestCleanup(f, config)
+				}
+			}()
+			pod := framework.StartVolumeServer(cs, config)
+			serverIP := pod.Status.PodIP
+			framework.Logf("Ceph server IP address: %v", serverIP)
 			By("sleeping a bit to give ceph server time to initialize")
 			time.Sleep(20 * time.Second)
 
@@ -286,13 +389,15 @@ var _ = framework.KubeDescribe("Volumes", func() {
 			}
 
 			defer func() {
-				if err := cs.CoreV1().Secrets(namespace.Name).Delete(secret.Name, nil); err != nil {
-					framework.Failf("unable to delete secret %v: %v", secret.Name, err)
+				if clean {
+					if err := cs.Core().Secrets(namespace.Name).Delete(secret.Name, nil); err != nil {
+						framework.Failf("unable to delete secret %v: %v", secret.Name, err)
+					}
 				}
 			}()
 
 			var err error
-			if secret, err = cs.CoreV1().Secrets(namespace.Name).Create(secret); err != nil {
+			if secret, err = cs.Core().Secrets(namespace.Name).Create(secret); err != nil {
 				framework.Failf("unable to create test secret %s: %v", secret.Name, err)
 			}
 
@@ -323,7 +428,8 @@ var _ = framework.KubeDescribe("Volumes", func() {
 	// (/usr/bin/nova, /usr/bin/cinder and /usr/bin/keystone)
 	// and that the usual OpenStack authentication env. variables are set
 	// (OS_USERNAME, OS_PASSWORD, OS_TENANT_NAME at least).
-	Describe("Cinder [Feature:Volumes]", func() {
+
+	framework.KubeDescribe("Cinder [Feature:Volumes]", func() {
 		It("should be mountable", func() {
 			framework.SkipUnlessProviderIs("openstack")
 			config := framework.VolumeTestConfig{
@@ -339,7 +445,11 @@ var _ = framework.KubeDescribe("Volumes", func() {
 			framework.Logf("cinder output:\n%s", outputString)
 			Expect(err).NotTo(HaveOccurred())
 
-			defer DeleteCinderVolume(volumeName)
+			defer func() {
+				// Ignore any cleanup errors, there is not much we can do about
+				// them. They were already logged.
+				DeleteCinderVolume(volumeName)
+			}()
 
 			// Parse 'id'' from stdout. Expected format:
 			// |     attachments     |                  []                  |
@@ -362,8 +472,10 @@ var _ = framework.KubeDescribe("Volumes", func() {
 			Expect(volumeID).NotTo(Equal(""))
 
 			defer func() {
-				framework.Logf("Running volumeTestCleanup")
-				framework.VolumeTestCleanup(f, config)
+				if clean {
+					framework.Logf("Running volumeTestCleanup")
+					framework.VolumeTestCleanup(f, config)
+				}
 			}()
 
 			tests := []framework.VolumeTest{
@@ -392,47 +504,73 @@ var _ = framework.KubeDescribe("Volumes", func() {
 	////////////////////////////////////////////////////////////////////////
 	// GCE PD
 	////////////////////////////////////////////////////////////////////////
-	Describe("PD", func() {
-		var config framework.VolumeTestConfig
 
-		BeforeEach(func() {
+	framework.KubeDescribe("PD", func() {
+		// Flaky issue: #43977
+		It("should be mountable [Flaky]", func() {
 			framework.SkipUnlessProviderIs("gce", "gke")
-			config = framework.VolumeTestConfig{
+			config := framework.VolumeTestConfig{
 				Namespace: namespace.Name,
 				Prefix:    "pd",
-				// PD will be created in framework.TestContext.CloudConfig.Zone zone,
-				// so pods should be also scheduled there.
-				NodeSelector: map[string]string{
-					kubeletapis.LabelZoneFailureDomain: framework.TestContext.CloudConfig.Zone,
+			}
+
+			By("creating a test gce pd volume")
+			volumeName, err := framework.CreatePDWithRetry()
+			Expect(err).NotTo(HaveOccurred())
+			defer func() {
+				// - Get NodeName from the pod spec to which the volume is mounted.
+				// - Force detach and delete.
+				pod, err := f.PodClient().Get(config.Prefix+"-client", metav1.GetOptions{})
+				Expect(err).NotTo(HaveOccurred(), "Failed getting pod %q.", config.Prefix+"-client")
+				detachAndDeletePDs(volumeName, []types.NodeName{types.NodeName(pod.Spec.NodeName)})
+			}()
+
+			defer func() {
+				if clean {
+					framework.Logf("Running volumeTestCleanup")
+					framework.VolumeTestCleanup(f, config)
+				}
+			}()
+
+			tests := []framework.VolumeTest{
+				{
+					Volume: v1.VolumeSource{
+						GCEPersistentDisk: &v1.GCEPersistentDiskVolumeSource{
+							PDName:   volumeName,
+							FSType:   "ext3",
+							ReadOnly: false,
+						},
+					},
+					File: "index.html",
+					// Randomize index.html to make sure we don't see the
+					// content from previous test runs.
+					ExpectedContent: "Hello from GCE from namespace " + volumeName,
 				},
 			}
-		})
 
-		It("should be mountable with ext3", func() {
-			testGCEPD(f, config, cs, "ext3")
-		})
-		It("should be mountable with ext4", func() {
-			testGCEPD(f, config, cs, "ext4")
-		})
-		It("should be mountable with xfs", func() {
-			// xfs is not supported on gci
-			// and not installed by default on debian
-			framework.SkipUnlessNodeOSDistroIs("ubuntu")
-			testGCEPD(f, config, cs, "xfs")
+			framework.InjectHtml(cs, config, tests[0].Volume, tests[0].ExpectedContent)
+
+			fsGroup := int64(1234)
+			framework.TestVolumeClient(cs, config, &fsGroup, tests)
 		})
 	})
 
 	////////////////////////////////////////////////////////////////////////
 	// ConfigMap
 	////////////////////////////////////////////////////////////////////////
-	Describe("ConfigMap", func() {
+
+	framework.KubeDescribe("ConfigMap", func() {
 		It("should be mountable", func() {
 			config := framework.VolumeTestConfig{
 				Namespace: namespace.Name,
 				Prefix:    "configmap",
 			}
 
-			defer framework.VolumeTestCleanup(f, config)
+			defer func() {
+				if clean {
+					framework.VolumeTestCleanup(f, config)
+				}
+			}()
 			configMap := &v1.ConfigMap{
 				TypeMeta: metav1.TypeMeta{
 					Kind:       "ConfigMap",
@@ -447,11 +585,11 @@ var _ = framework.KubeDescribe("Volumes", func() {
 					"third":  "this is the third file",
 				},
 			}
-			if _, err := cs.CoreV1().ConfigMaps(namespace.Name).Create(configMap); err != nil {
+			if _, err := cs.Core().ConfigMaps(namespace.Name).Create(configMap); err != nil {
 				framework.Failf("unable to create test configmap: %v", err)
 			}
 			defer func() {
-				_ = cs.CoreV1().ConfigMaps(namespace.Name).Delete(configMap.Name, nil)
+				_ = cs.Core().ConfigMaps(namespace.Name).Delete(configMap.Name, nil)
 			}()
 
 			// Test one ConfigMap mounted several times to test #28502
@@ -498,10 +636,13 @@ var _ = framework.KubeDescribe("Volumes", func() {
 	////////////////////////////////////////////////////////////////////////
 	// vSphere
 	////////////////////////////////////////////////////////////////////////
-	Describe("vsphere [Feature:Volumes]", func() {
+
+	framework.KubeDescribe("vsphere [Feature:Volumes]", func() {
 		It("should be mountable", func() {
 			framework.SkipUnlessProviderIs("vsphere")
-			var volumePath string
+			var (
+				volumePath string
+			)
 			config := framework.VolumeTestConfig{
 				Namespace: namespace.Name,
 				Prefix:    "vsphere",
@@ -518,8 +659,10 @@ var _ = framework.KubeDescribe("Volumes", func() {
 			}()
 
 			defer func() {
-				framework.Logf("Running volumeTestCleanup")
-				framework.VolumeTestCleanup(f, config)
+				if clean {
+					framework.Logf("Running volumeTestCleanup")
+					framework.VolumeTestCleanup(f, config)
+				}
 			}()
 
 			tests := []framework.VolumeTest{
@@ -543,11 +686,10 @@ var _ = framework.KubeDescribe("Volumes", func() {
 			framework.TestVolumeClient(cs, config, &fsGroup, tests)
 		})
 	})
-
 	////////////////////////////////////////////////////////////////////////
 	// Azure Disk
 	////////////////////////////////////////////////////////////////////////
-	Describe("Azure Disk [Feature:Volumes]", func() {
+	framework.KubeDescribe("Azure Disk [Feature:Volumes]", func() {
 		It("should be mountable [Slow]", func() {
 			framework.SkipUnlessProviderIs("azure")
 			config := framework.VolumeTestConfig{
@@ -563,8 +705,10 @@ var _ = framework.KubeDescribe("Volumes", func() {
 			}()
 
 			defer func() {
-				framework.Logf("Running volumeTestCleanup")
-				framework.VolumeTestCleanup(f, config)
+				if clean {
+					framework.Logf("Running volumeTestCleanup")
+					framework.VolumeTestCleanup(f, config)
+				}
 			}()
 			fsType := "ext4"
 			readOnly := false
@@ -593,42 +737,3 @@ var _ = framework.KubeDescribe("Volumes", func() {
 		})
 	})
 })
-
-func testGCEPD(f *framework.Framework, config framework.VolumeTestConfig, cs clientset.Interface, fs string) {
-	By("creating a test gce pd volume")
-	volumeName, err := framework.CreatePDWithRetry()
-	Expect(err).NotTo(HaveOccurred())
-	defer func() {
-		// - Get NodeName from the pod spec to which the volume is mounted.
-		// - Force detach and delete.
-		pod, err := f.PodClient().Get(config.Prefix+"-client", metav1.GetOptions{})
-		Expect(err).NotTo(HaveOccurred(), "Failed getting pod %q.", config.Prefix+"-client")
-		detachAndDeletePDs(volumeName, []types.NodeName{types.NodeName(pod.Spec.NodeName)})
-	}()
-
-	defer func() {
-		framework.Logf("Running volumeTestCleanup")
-		framework.VolumeTestCleanup(f, config)
-	}()
-
-	tests := []framework.VolumeTest{
-		{
-			Volume: v1.VolumeSource{
-				GCEPersistentDisk: &v1.GCEPersistentDiskVolumeSource{
-					PDName:   volumeName,
-					FSType:   fs,
-					ReadOnly: false,
-				},
-			},
-			File: "index.html",
-			// Randomize index.html to make sure we don't see the
-			// content from previous test runs.
-			ExpectedContent: "Hello from GCE from namespace " + volumeName,
-		},
-	}
-
-	framework.InjectHtml(cs, config, tests[0].Volume, tests[0].ExpectedContent)
-
-	fsGroup := int64(1234)
-	framework.TestVolumeClient(cs, config, &fsGroup, tests)
-}
